Remove version pinning of apt packages in Dockerfile
Problem:
Dockerfile no longer builds because the exact versions
of the packages that are installing using apt in the Dockerfile
are no longer available.  The add-on is built using a specific
version of the base Debian image.  We don't need to pin
versions since that specific version of the Debian image should
use compatible libraries.

Implementation:
Remove the specific versions of apt packages to be installed.